backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client for OAuth
httpx==0.26.0

# OpenAI
openai==1.12.0

# Environment and settings
python-dotenv==1.0.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Development
ruff==0.2.1
pytest==7.4.4
pytest-asyncio==0.23.4

# CORS
python-multipart==0.0.9
